+++ templates/basic-template/README.md
@@ -1,0 +1,1 @@
+Congratulations on the first day of the rest of your life. Hold onto your pants cuz your mind is about to be blown
