+++ templates/basic-template/program/src/lib.rs
@@ -9,6 +9,9 @@ use alloc::{string::ToString, vec::Vec};
 use entropy_programs_core::{bindgen::Error, bindgen::*, export_program, prelude::*};
 use serde::{Deserialize, Serialize};
 
+#[cfg(test)]
+mod tests;
+
 // TODO confirm this isn't an issue for audit
 register_custom_getrandom!(always_fail);
 
@@ -46,28 +49,3 @@ impl Program for {{project-name | upper_camel_case}} {
 }
 
 export_program!({{project-name | upper_camel_case}});
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_should_sign() {
-        let signature_request = SignatureRequest {
-            message: b"some_message".to_vec(),
-            auxilary_data: None,
-        };
-
-        assert!({{project-name | upper_camel_case}}::evaluate(signature_request, None).is_ok());
-    }
-
-    #[test]
-    fn test_should_fail() {
-        let signature_request = SignatureRequest {
-            message: Vec::new(),
-            auxilary_data: None,
-        };
-
-        assert!({{project-name | upper_camel_case}}::evaluate(signature_request, None).is_err());
-    }
-}
